Propose new boot order to users
Propose new boot order to owners of VMs that exist on nodes
running Proxmox version 6.3 or greater.

Instead of returning the legacy boot order to the UI, a new boot
order is created and sent instead. The new order mimics the old
order, but is compliant with the new format. The new order will
take effect only if the user explicitly updates.