Update UI boot order selection element
Update the UI boot order selection element. It has been modified to
support the new 'order=' format for the Proxmox boot order option
in a more intuitive way for the user. The new element allows users
to enable any bootable device and choose the order for these
devices to boot in.

The new UI element is a table with four columns. The first column
indicates the order of the corresponding device. The second column
indicates whether the device is enabled as bootable. The third
column contains the device name. The fourth column contains the
device description. The order of the columns is changed by
dragging, enabled by SortableJS.

In cases of legacy nodes, the older dropdown boxes are used.